type of timeout was changed, getQueries util was created, .env file with API was created
