+++ src/store/data/allProducts/asyncAction.ts
@@ -3,19 +3,20 @@ import axios from 'axios';
 
 import { IProductData } from '@/models/models';
 import { IFetchParams } from '@store/data/allProducts/types';
+import { getQueriesForProducts } from '@/utils/queriesForUrl/getQueriesForProducts/getQueriesForProducts';
 
 export const fetchAllProducts = createAsyncThunk(
   'products/fetchAllProducts',
   async ({ page, size, sort, category, query }: IFetchParams) => {
-    const queryParams = [];
-    if (page !== undefined && page !== null) queryParams.push(`page=${page}`);
-    if (size) queryParams.push(`size=${size}`);
-    if (sort) queryParams.push(`sort=${sort}`);
-    if (category) queryParams.push(`category.name=${category}`);
-    if (query) queryParams.push(`${query}`);
+    const queryString = getQueriesForProducts({
+      page,
+      size,
+      sort,
+      category,
+      query,
+    });
 
-    const queryString = queryParams.join('&');
-    const url = `http://localhost:8080/products?${queryString}`;
+    const url = `${process.env.REACT_APP_API_URL}/products?${queryString}`;
 
     const { data } = await axios.get<IProductData>(url);
