+++ src/store/data/reviews/asyncAction.ts
@@ -7,7 +7,7 @@ export const fetchReviews = createAsyncThunk(
   'reviews/fetchReviews',
   async () => {
     const { data } = await axios.get<IProductData>(
-      `http://localhost:8080/reviews/best-rate`,
+      `${process.env.REACT_APP_API_URL}/reviews/best-rate`,
     );
 
     return data;
